Use sys.exit() rather than exit()
exit() is not guaranteed to exist in non-interactive Python, and sys.exit()
is easier to replace with a mock for testing.

Ideally these exit calls should be replaced with a clean normal exit (i.e.
a simple return from main()) or an exception. Calling exit() makes life much
more difficult when trying to use this code from elsewhere.