Add functionality to read in a Relion pipeline and get job timing information from log files (#67)
Relion provides a `default_pipeline.star` file containing the scheduled Relion jobs and their inputs as well as some of their outputs. Edges between these nodes are provided (links between job outputs and the next jobs in the pipeline). A graph can then be constructed containing this information. Log files produced by, for example, `cryolo_relion_it.py` give information on when the jobs were run which supplements the pipeline information. This PR adds a few classes to contain this information and provides some basic visualisation of the pipeline through the `graphviz` package. This is currently added as a dev requirement as it is used in some tests but should not be required for the majority of the functionality.

The added classes are:
* `ProcessNode` which has the file or directory path within the Relion project as an attribute, along with the `ProcessNode`s it acts as an input for. It also has an `attributes` dictionary to store further information such as start and end times of a job
* `ProcessGraph` behaves as a list of `ProcessNode`s. Contains methods to: add or remove `ProcessNode`s dealing with the edges in the graph appropriately, split itself into connected subgraphs, merge with another `ProcessGraph`, and display itself using `graphviz`
* `RelionPipeline` has a couple of different `ProcessGraphs` as attributes. These are the full Relion pipeline (`_nodes`) and the compressed version that just has the jobs as nodes (`_job_nodes`), dropping the details of which exact files act as inputs to other jobs. Contains methods that collect timing information from log files and add them to the relevant `ProcessNode` `attributes`

`Project` then contains a `RelionPipeline` which can be loaded through the `_load_pipeline` method. In the future it is probably better to have `Project` inherit from `RelionPipeline` and add extra functionality such as results collection.

A small command line interface to the node displaying through `graphviz` is also provided for convenience.