Make sure postprocessing jobs are not stopped immediately which was causing the wrapper to exit early (#88)
In #85 a timeout was applied that stopped preprocessing jobs if the last image file is more than 30 minutes old. In the case where processing is done on an old data collection this leads to an earlier exit than is desirable as the preprocessing jobs are stopped quickly and the wrapper may exit without results collection. Here it is required that `cryolo_relion_it` has been running for 10 minutes before the preprocessing jobs can be stopped.

Some changes are also made to reduce the number of times the `default_pipeline.star` file is accessed (which should remove a race condition arising from the fact that we sometimes want those calls to be reading the same star file data) and to stop `_calculate_relative_job_times` from crashing if there are no registered jobs.