Stop Relion processing if there are no currently running jobs and no jobs have been submitted for 10 minutes (#92)
If there are no jobs currently running ("running jobs" are here taken to be jobs that have started according to the log files but not posted a success or failure) and no jobs were submitted in the last 10 minutes then delete the `RUNNING_RELION_IT` file to stop the pipeline. This then allows the wrapper to exit. Before this change the wrapper waited for `cryolo_relion_it` to stop running which may rely on the manual deletion of the RUNNING file.