Check that all imported files have been preprocessed before allowing an exit (#101)
Check that all the files in `Import/job001/movies.star` have been through a MotionCorr job by checking the file names against the cached MotionCorr micrograph names (after removing the file extension as it is `tiff` for the Import files and `mrc` for the MotionCorr results). If this check passes then it is only checked again if the list of imported files changes. 

Preprocessing jobs are checked to ensure that they have all been run more recently than the last MotionCorr job that finished the complete set of imported files.