Update the ISPyB messages sent for MotionCorrection and CtfFind Relion stages and add messages for other stages (#98)
The current ISPyB messages for the MotionCorrection and CtfFind Relion stages do not contain records for the columns relating to the options with which Relion was run. These can be added in using the `RelionItOptions` object. In the future we will also need ISPyB messages for further stages of the Relion pipeline. The necessary `JobType` classes already exist for 2D and 3D classification. Small `JobType` classes are added here for auto-picking stages: either Relion's own AutoPick or crYOLO along with the relevant ISPyB commands (and the commands for the classification stages).

Co-authored-by: Markus Gerstel <2102431+Anthchirp@users.noreply.github.com>