Add Icebreaker external jobs and histogram (#106)
Add three external Icebreaker jobs to run at different stages in the Relion pipeline. 
The output from one of these jobs is used to make histograms of ice thickness. These are sent to ISPyB as autoProcProgramAttachments in .pdf and .json format. 
The output from the other jobs is stored in Icebreaker directories in the Relion working dir.

Co-authored-by: Markus Gerstel <markus.gerstel@diamond.ac.uk>
Co-authored-by: Daniel Hatton <daniel.hatton@diamond.ac.uk>