Use DLSImages service to convert from mrc to jpeg for CTF diagnostic image and motion corrected micrograph (#107)
To display Relion output images in SynchWeb we need to convert them from mrc format to jpeg and then insert the path into ISPyB. These changes should provide the infrastructure for doing that (on the wrapper side) for CTF diagnostic plots and motion corrected micrographs. 

The conversion will be done by the DLSImages zocalo service using the `images_service_plugin` provided here. This will require changes to the implementation of the Images service. Sending messages to the service will require a change to the zocalo recipe as well.

Co-authored-by: Markus Gerstel <markus.gerstel@diamond.ac.uk>