Add Table class to support modelling of ISPyB database entries from within the wrapper (#111)
Propose adding a `Table` class to which rows can be added with automatically incremented primary keys to produce a DB model for use in conjunction with the Zocalo buffer table. Allow for certain constraints to be placed on added rows, such as uniqueness of values for certain columns, or auto-incrementing columns. Such constraints can be used to stop duplicate entries being created in an environment where results will be collected regularly from the same set of jobs. 

There are two single dispatches included (`insert` and `construct_message`) which are intended for use in functionality to be added later.

Some documentation included. Can be better integrated with existing documentation later if merged.

Co-authored-by: Markus Gerstel <2102431+Anthchirp@users.noreply.github.com>
Co-authored-by: Markus Gerstel <markus.gerstel@diamond.ac.uk>