Add `Node` and `Graph` classes to allow organisation of data collection and database insertion (#113)
There are already `ProcessNode` and `ProcessGraph` classes to mimic the structure of a Relion project and help keep track of which jobs have run/been scheduled. `DBNode` and `DBGraph` are added in order to get data collected from the `ProcessNode`s to the correct tables in the database model introduced in #111. Refactor to have both types of node inherit from a base class to allow data transfer between the two. The project should then be able to be described as a graph structure, one part of which follow the Relion job structure, which is linked to another part that mimics the structure of the database schema.

Co-authored-by: Markus Gerstel <2102431+Anthchirp@users.noreply.github.com>