Use a model of the database to collect results from within the wrapper (#115)
Tie components from #112 and #113 into the actual collection of results and construction of messages to send to the ISPyB service done by the wrapper. Keeping track of results that have already been sent is then handled by the database model infrastructure and replaces the component that did this before. This should allow the use of the Zocalo buffer table on the ISPyB service side and will hopefully be more extensible in case of future table additions. ISPyB service messages are also changed here to reflect the Zocalo buffer functionality recently added to the service.