Fix ISPyB messages for classification tables (#116)
Every time the project was loaded two undesirable things happened:

- the `_data_pipeline`, which contains information about which jobs are inputs for which tables, would be appended to and grow in size
- the database nodes in the pipeline are persistent across loads and so would get new nodes representing the jobs appended to their `_in` list

Fixing these problems also fixes aberrant behaviour in the ISPyB message generation for the classification tables when running in real time. 

In addition the code now handles the case where multiple nodes want to pass lists to the same node. This happens with the tutorial dataset for example where there are multiple auto picking jobs which all need to connect to the `ParticlePicker` table.