Add image service plugin to make particle picker images (#118)
The plugin requires a background image, for which we use the motion corrected jpegs that should already exist, and a set of particle coordinates (as well as information on the diameter). Particle coordinates are collected from files and are cached to avoid repeated file scraping in a manner analogous to what is already done for motion correction drift data.