Add movie creation time stamps to ISPyB service message (#130)
When collecting data from the motion correction job additionally collect creation time stamps for the movies that were motion corrected. The path to the imported movie can be found in the same star file as the drift data which is already collected. However, it is in a key-value section rather than a loopy one so need to extend parsing capabilities marginally. This should resolve #129. A small change will also be needed in the ISPyB service to make use of the provided time stamp.