Check preprocessing status if there are no new batches (#145)
If there haven't been any new batches then check that the background process running preprocessing is still alive. If it has exited then finish the thread running Relion. If it exited with a failing error code then remove the file used to stop `relion_it` running. The wrapper currently deletes this file if it is still there at the end of processing, which it should be. We can return a failure state if it isn't as it must have been deleted because of the above preprocessing failure (`relion_it` doesn't remove it anywhere else).