Add a first attempt Relion 4 pipeline (#149)
The Relion 4 pipeline should behave the same as the current Relion 3.1 pipeline. Makes use of the CCP-EM pipeliner as an API to run Relion jobs. Adds an option to the run_pipeline command line tool to choose the Relion version of the pipeline (choices are 3.1 or 4.0). Code is a bit messy at the moment as a result of

- Complicated control flow, particularly when considering a second pass
- Methods for determining the best class which can be refactored later

There will probably be some substantial refactoring in the future