Calculate initial model class number from occupation (#176)
Currently the initial model class number is collected from the name of the file as determined from `default_pipeline.star`. When using `initial_model.mrc` this can't be done so instead parse the data star file of the initial model job and choose the largest class (not sure if this is strictly correct).