Add particle diameter option (#180)
Convert `RelionItOptions` to a `pydantic.BaseModel` in order to utilise the pydantic `root_validator` to calculate extraction box sizes and classification mask diameters according to a provided particle diameter. If the particle diameter is changed the dependencies are recalculated. If particle_diameter is left as 0 then these calculations are not performed.

---------

Co-authored-by: Stephen Riggs <122790971+stephen-riggs@users.noreply.github.com>