Run a single 3D classification batch (#184)
When automated 2D class selection is being used just run a single batch of 3D classification. Allow the size of this batch to be varied independently of the 2D classification batch size. Every time a 2D class batch is complete re-run the 3D classification job with the new data until the 3D batch is complete. This will reduce the number of 3D class jobs and will hopefully be more useful

---------

Co-authored-by: stephen-riggs <stephen.riggs@diamond.ac.uk>